1. Fix x86-64 codegen for binary operators with 64-bit imm values
2. When decoding floating point instructions on x86-64, the address size
override prefix means 64-bit address size rather than 16-bit

3. Fix a linking problem of examples

4. Do not use negative values when reading jump tables

+++ dyninstAPI/src/inst-aarch64.C
@@ -971,7 +971,7 @@ void registerSpace::saveClobberInfo(const instPoint *location)
 #endif
 
 
-bool doNotOverflow(int value) {
+bool doNotOverflow(int64_t value) {
     // This function seems irrelevant for aarch64 two reasons:
     // 1) it's used only once in Operator AST node and
     // 2) when value is passed to emitImm, it ends in a moveValueToReg or it's
